+++ src/routes/SetUp/Products/index.js
@@ -180,17 +180,15 @@ class Products extends Component {
   };
 
   onDeletePopUp = (recordId) => {
-    return (
-      <Popconfirm
-        title={<IntlMessages id="products.message.delete"/>}
-        onConfirm={() => {
-          this.props.onDeleteProduct({ids: [recordId]}, this);
-        }}
-        okText={<IntlMessages id="common.yes"/>}
-        cancelText={<IntlMessages id="common.no"/>}>
-        <i className="icon icon-trash"/>
-      </Popconfirm>
-    )
+    const {messages} = this.props.intl;
+    confirm({
+      title: messages["products.message.delete"],
+      okText: messages["common.yes"],
+      cancelText: messages["common.no"],
+      onOk: () => {
+        this.props.onDeleteProduct({ids: [recordId]}, this);
+      }
+    })
   };
 
   onShowItemOptions = () => {
